Deque is finished for now.
It uses an array as the internal structure. This gives the following
advantages.
We can mitigate the need to allocate after every push, and deallocate
after every pop.
We gain the reusability of nodes this way

We can initialize the queue to any arbitrary size using the
hkl_deque_new_sized.
This means that if given a relatively ballparked estimate for the memory
needs of a queue
we are able to minimize the calls to malloc to 1 or slightly more.

We are able to use the deque to implement internal arrays in our
langague. They are dynamically sized and therefore will allow HKL to use
them as much as needed.

We can locate the nth item in the array given an index such that -size
<= index < size.

We can locate the nth item in the array given that the queue isn't
empty. This means that cyclical arrays are now as easy to implement as a
single for loop with no extra work required.
enter the commit message for your changes. Lines starting

FIXED Issues with not compiling on warnings.

+++ src/hkl_deque.c
@@ -176,12 +176,12 @@ void *hkl_deque_findn(HklDeque *deque, int n)
   assert(deque != NULL);
   
   int *data = NULL;
-  if(n >= 0 && n < deque->size)
+  if(n >= 0 && n < (int)deque->size)
   {
     int location = (deque->front + n) % deque->max;
     data = deque->queue[location];
   }
-  else if(n < 0 && -n < deque->size + 1)
+  else if(n < 0 && -n < (int)deque->size + 1)
   {
     //we want to offset the value so that -1 is the back
     int location = (deque->back + (n + 1)) % deque->max;  
@@ -196,7 +196,7 @@ void *hkl_deque_findn_cyclical(HklDeque *deque, int n)
   assert(deque != NULL);
   int *data = NULL;
 
-  if(deque->size > 0)//only null on an empty queue
+  if((int)deque->size > 0)//only null on an empty queue
   {
     int index = (n % deque->size) + deque->front;//get a cyclical index 
     data = deque->queue[index];
